+++ packages/reactabular/CHANGELOG.md
@@ -1,3 +1,10 @@
+5.0.0 / 2016-09-28
+==================
+
+## reactabular-easy
+
+  * Breaking - Push children toggle behavior to `onToggleShowingChildren` prop. See the readme for a sample implementation. This makes the implementation more flexible. You can trigger `reactabular-tree` `collapseAll` and `expandAll` over `showingChildren` for instance.
+
 4.3.0 / 2016-09-27
 ==================